# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library",
        "description": "A php library website",
        "contact": {
            "name": "Lucy",
            "email": "alan8365@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Develop environment"
        }
    ],
    "paths": {
        "/book": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "List all posts by paginate",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list success"
                    }
                }
            }
        },
        "/book/{isbn}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Show book detail",
                "operationId": "App\\Http\\Controllers\\Api\\BookController::detail",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get detail success."
                    },
                    "404": {
                        "description": "Book not found."
                    }
                }
            }
        },
        "/book/favorite": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Show all favorite book.",
                "operationId": "App\\Http\\Controllers\\Api\\BookController::listFavorite",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get favorite book success."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/book/favorite/{isbn}": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Chage book favorite status.",
                "operationId": "App\\Http\\Controllers\\Api\\BookController::setFavorite",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Change book favorite status success."
                    },
                    "404": {
                        "description": "Book not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/book/search": {
            "get": {
                "tags": [
                    "Book"
                ],
                "summary": "Search book by title",
                "operationId": "App\\Http\\Controllers\\Api\\BookController::search",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/dashboard/book": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "List all book",
                "operationId": "App\\Http\\Controllers\\Api\\DashboardController::getAllBook",
                "responses": {
                    "200": {
                        "description": "list success"
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Create new book.",
                "operationId": "App\\Http\\Controllers\\Api\\DashboardController::storeBook",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "publisher",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "publication_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "summary",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img_src",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create success"
                    },
                    "400": {
                        "description": "Field invalid."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/dashboard/book/{isbn}": {
            "put": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Update target book.",
                "operationId": "App\\Http\\Controllers\\Api\\DashboardController::updateBook",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "publisher",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "publication_date",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "summary",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "img_src",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update success"
                    },
                    "400": {
                        "description": "Field invalid."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Remove target book.",
                "operationId": "App\\Http\\Controllers\\Api\\DashboardController::removeBook",
                "parameters": [
                    {
                        "name": "isbn",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete success"
                    },
                    "401": {
                        "description": "Unauthorized."
                    },
                    "404": {
                        "description": "Book not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/store": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user.",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::store",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Register success"
                    },
                    "400": {
                        "description": "Field invalid"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get token in website.",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List success"
                    },
                    "400": {
                        "description": "Field invalid"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/auth/whoAmI": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get current user info.",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::whoAmI",
                "responses": {
                    "200": {
                        "description": "Get user info success."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cancel token.",
                "operationId": "App\\Http\\Controllers\\Api\\UserController::logout",
                "responses": {
                    "200": {
                        "description": "Logout success."
                    },
                    "401": {
                        "description": "Unauthorized."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth collection"
        },
        {
            "name": "Book",
            "description": "Book collection"
        },
        {
            "name": "Dashboard",
            "description": "Dashboard collection"
        }
    ]
}